requirements.txt
safetensors>=0.3.0
git+https://github.com/huggingface/diffusers.git
gradio_logsview@https://huggingface.co/spaces/cocktailpeanut/gradio_logsview/resolve/main/gradio_logsview-0.0.17-py3-none-any.whl
transformers>=4.30.0
lycoris-lora==1.8.3
flatten_json
pyyaml
oyaml
tensorboard
kornia
invisible-watermark
einops
accelerate>=0.21.0
toml
albumentations
pydantic
omegaconf
k-diffusion
open_clip_torch
timm
prodigyopt
controlnet_aux==0.0.7
python-dotenv
bitsandbytes
hf_transfer
lpips
pytorch_fid
optimum-quanto
sentencepiece
huggingface_hub>=0.16.0
peft
gradio>=4.0.0
python-slugify>=6.0.0
imagesize
google-generativeai>=0.3.0
datasets>=2.14.0
